% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pogit}
\alias{print.pogit}
\title{Print an object of class \code{pogit}}
\usage{
\method{print}{pogit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pogit}}

\item{...}{further arguments passed to or from other methods (not used)}
}
\description{
The default print method for a \code{pogit} object.
}
\details{
Returns basic information about the model, the number of observations and 
covariates used, the number of regression effects subject to selection, 
MCMC options and the runtime used for the sampling algorithm. See
\code{\link{summary.pogit}} for more details.
}
\author{
Michaela Dvorzak <m.dvorzak@gmx.at>
}
